api/requirements.txt
fastapi==0.121.3
pydantic==2.12.4
python-dotenv==1.2.1
requests==2.32.5
mangum==0.18.0
uvicorn[standard]==0.38.0
